add jwt token

+++ part4/bloglist/controllers/blogsRouter.js
@@ -1,9 +1,21 @@
 const blogsRouter = require('express').Router()
+const config = require('../utils/config')
 const Blog = require('../models/models')
+const User = require('../models/user')
+const jwt = require('jsonwebtoken')
+
+const getTokenFrom = request => {
+  const authorization = request.get('authorization')
+  if (authorization && authorization.startsWith('Bearer ')) {
+    return authorization.replace('Bearer ', '')
+  }
+  return null
+}
 
 blogsRouter.get('/', async (request, response, next) => {
   try {
-    const blogs = await Blog.find({})
+    const blogs = await Blog
+      .find({}).populate('user', { username: 1, name: 1 })
     response.json(blogs)
   } catch (exception) {
     next(exception)
@@ -11,6 +23,22 @@ blogsRouter.get('/', async (request, response, next) => {
 })
 
 blogsRouter.delete('/:id', async (request, response, next) => {
+  const blogToDelete = await Blog.findById(request.params.id)
+
+  if (!request.user) {
+    return response.status(401).json({
+      error: 'user token missing or invalid user'
+    })
+  }
+
+  const user = request.user
+
+  if (blogToDelete.user.toString() !== user._id.toString()) {
+    return response.status(401).json({
+      error: 'only users who wrote blog entry can delete entry'
+    })
+  }
+  
   try {
     await Blog.findByIdAndRemove(request.params.id)
     response.status(204).end()
@@ -21,6 +49,12 @@ blogsRouter.delete('/:id', async (request, response, next) => {
 
 blogsRouter.post('/', async (request, response, next) => {
   const body = request.body
+
+  const decodedToken = jwt.verify(getTokenFrom(request), config.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+  const user = await User.findById(body.userId)
 
   if (!body.title || !body.author || !body.url) {
     return response.status(400).json({
@@ -34,11 +68,14 @@ blogsRouter.post('/', async (request, response, next) => {
     title: body.title,
     author: body.author,
     url: body.url,
-    likes: body.likes
+    likes: body.likes,
+    user: user._id
   })
 
   try {
     const savedBlog = await blog.save()
+    user.blogs = user.blogs.concat(savedBlog._id)
+    await user.save()
     response.status(201).json(savedBlog)
   } catch (exception) {
     next(exception)
